refactor: extract OperationStatus to separate status.go file
Move OperationStatus enum from results.go to its own file as it's a
universal type used across handlers and operations. This is the first
step in reorganizing pkg/types to contain only truly universal types.

Part of #704
