+++ pkg/types/results.go
@@ -5,17 +5,6 @@ import (
 	"time"
 )
 
-// OperationStatus defines the state of an operation execution
-type OperationStatus string
-
-const (
-	StatusReady    OperationStatus = "ready"
-	StatusSkipped  OperationStatus = "skipped"
-	StatusConflict OperationStatus = "conflict"
-	StatusError    OperationStatus = "error"
-	StatusUnknown  OperationStatus = "unknown"
-)
-
 // DisplayResult is the top-level structure for commands that produce rich output.
 // This replaces the old PackStatusResult and is used by status, deploy, and install commands.
 type DisplayResult struct {
